Move process_gcm_data to climate_prepro
